refactor: create hooks for listeners
